(refactor and ui update mode) : remove hardcoded strings from state screens + refactor
