+++ app/src/main/java/com/kev/lydia/ui/list/components/state/LoaderScreen.kt
@@ -1,4 +1,4 @@
-package com.kev.lydia.ui
+package com.kev.lydia.ui.list.components.state
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
@@ -7,7 +7,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
-// ------------------------- Loader -------------------------
 @Composable
 fun LoaderScreen() {
     Box(
